requirements.txt
alembic==1.14.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==4.8.0
apispec==6.8.1
app==0.0.1
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
filestorage==0.0.6
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==5.6.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
loginform==1.2.0
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.1
mimesis==18.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.1.3
WTForms==3.1.2
